Add files via upload
loaded, cleaned, and feature engineered the Churn dataset. We checked for outliers using IQR method and creating a boxplot but no major outliers exist. However few incorrect records exist where tenure is zero and monthly charges is some positive value so decided to remove those records. The "total value" column also had missing values for same records so have deleted these 11 records from the data set. 

To do some more data cleaning checked on the extra columns and as customerID is not a deciding factor for Churn rate so removed the column as of now. 

Converted all the categorical columns to numeric columns with binary label encoding. Now all the data columns are numeric. Also created a new feature, the TotalCharge/tenure ratio which is a monthly charge for a particular customer. Then created a plot and histogram for this new feature and it is observed that customers with less total charge and tenure ratio tend to churn less. 

Created an EDA with the updated and attached the same.
Finally, saved the updated data as a CSV to use for further analysis.